@@ -1,38 +1,33 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <LangVersion>8</LangVersion>
-    </PropertyGroup>
-
-    <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-      <DocumentationFile>bin\Debug\plodo.Backend.API.xml</DocumentationFile>
-    </PropertyGroup>
-
-    <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-      <DocumentationFile>bin\Release\plodo.Backend.API.xml</DocumentationFile>
+        <TargetFramework>net8.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+        <LangVersion>12</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-        <FrameworkReference Include="Microsoft.AspNetCore.App" />
-    </ItemGroup>
-    
-    <ItemGroup>
-        <PackageReference Include="Lamar.Microsoft.DependencyInjection" Version="4.1.5" />
-        <PackageReference Include="Lib.AspNetCore.ServerSentEvents" Version="4.0.0" />
-        <PackageReference Include="Microsoft.ApplicationInsights" Version="2.13.1" />
-        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.13.1" />
-        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.0.3" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="4.0.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="4.0.0" />
-        <PackageReference Include="Microsoft.OpenApi" Version="1.1.4" />
-        <PackageReference Include="Serilog" Version="2.8.0" />
-        <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
+        <PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0" />
+        <PackageReference Include="Lamar" Version="13.1.0" />
+        <PackageReference Include="Lamar.Microsoft.DependencyInjection" Version="13.1.0" />
+        <PackageReference Include="Lib.AspNetCore.ServerSentEvents" Version="9.0.0" />
+        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.22.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.10" />
+        <PackageReference Include="Serilog.AspNetCore" Version="8.0.2" />
         <PackageReference Include="Serilog.Enrichers.AspNetCore.HttpContext" Version="1.0.1" />
-        <PackageReference Include="Serilog.Sinks.ApplicationInsights" Version="3.1.0" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.1.0" />
-        <PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="5.0.2" />
-        <PackageReference Include="YamlDotNet" Version="8.1.0" />
+        <PackageReference Include="Serilog.Sinks.ApplicationInsights" Version="4.0.0" />
+        <PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.8.1" />
+        <PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="8.0.2" />
+        <PackageReference Include="YamlDotNet" Version="16.1.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Content Include="..\.dockerignore">
+        <Link>.dockerignore</Link>
+      </Content>
     </ItemGroup>
 
     <ItemGroup>
