@@ -1,16 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <LangVersion>8</LangVersion>
+        <TargetFramework>net8.0</TargetFramework>
+        <LangVersion>12</LangVersion>
+        <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Lamar" Version="4.2.1" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.1.2" />
-      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.6.0" />
-      <PackageReference Include="Nanoid" Version="2.1.0" />
-      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.6.0" />
+      <PackageReference Include="Lamar" Version="13.1.0" />
+      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="8.0.0" />
+      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="8.1.1" />
+      <PackageReference Include="Nanoid" Version="3.1.0" />
+      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.1.1" />
     </ItemGroup>
 
     <ItemGroup>
