@@ -1,32 +1,28 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 using Lamar;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 
-namespace plodo.Backend.API.Health
-{
-    public class ContainerHealthCheck : IHealthCheck
-    {
-        private readonly IContainer _ioc;
+namespace plodo.Backend.API.Health;
 
-        public ContainerHealthCheck(IContainer container)
-        {
-            _ioc = container;
-        }
+public class ContainerHealthCheck : IHealthCheck
+{
+    private readonly IContainer _ioc;
+
+    public ContainerHealthCheck(IContainer container)
+    {
+        _ioc = container;
+    }
         
-        public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = new CancellationToken())
+    public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = new CancellationToken())
+    {
+        try
         {
-            try
-            {
-                _ioc.AssertConfigurationIsValid(AssertMode.ConfigOnly);
-                return HealthCheckResult.Healthy();
-            }
-            catch (Exception e)
-            {
-                return HealthCheckResult.Unhealthy("Container configuration is not valid", e);
-            }
-            
+            _ioc.AssertConfigurationIsValid(AssertMode.ConfigOnly);
+            return HealthCheckResult.Healthy();
         }
+        catch (Exception e)
+        {
+            return HealthCheckResult.Unhealthy("Container configuration is not valid", e);
+        }
+            
     }
 }
