@@ -1,14 +1,12 @@
-using System.Collections.Generic;
 using plodo.Backend.API.Validation;
 
-namespace plodo.Backend.API.Models
+namespace plodo.Backend.API.Models;
+
+public class CreateSessionRequest
 {
-    public class CreateSessionRequest
-    {
-        /// <summary>
-        /// A list of valid votes. Votes must be valid Unicode emojis
-        /// </summary>
-        [MultiVoteOptionValidator]
-        public IEnumerable<string> VotingOptions { get; set; }
-    }
+    /// <summary>
+    /// A list of valid votes. Votes must be valid Unicode emojis
+    /// </summary>
+    [MultiVoteOptionValidator]
+    public IEnumerable<string> VotingOptions { get; set; }
 }
