@@ -1,23 +1,20 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Text.RegularExpressions;
 
-namespace plodo.Backend.API.Validation
+namespace plodo.Backend.API.Validation;
+
+public class MultiVoteOptionValidator : ValidationAttribute
 {
-    public class MultiVoteOptionValidator : ValidationAttribute
+    public override bool IsValid(object value)
     {
-        public override bool IsValid(object value)
-        {
-            return value is IEnumerable<string> list && list.All(x => Regex.IsMatch(x, @"^[^\u0000-\u007F]+$"));
-        }
+        return value is IEnumerable<string> list && list.All(x => Regex.IsMatch(x, @"^[^\u0000-\u007F]+$"));
     }
+}
     
-    public class SingleVoteOptionValidator : ValidationAttribute
+public class SingleVoteOptionValidator : ValidationAttribute
+{
+    public override bool IsValid(object value)
     {
-        public override bool IsValid(object value)
-        {
-            return value is string str && Regex.IsMatch(str, @"^[^\u0000-\u007F]+$");
-        }
+        return value is string str && Regex.IsMatch(str, @"^[^\u0000-\u007F]+$");
     }
 }
