@@ -1,17 +1,16 @@
 ﻿using Cassandra;
 using Cassandra.Mapping;
 
-namespace plodo.Backend.Repositories
+namespace plodo.Backend.Repositories;
+
+public class RepositoryBase
 {
-    public class RepositoryBase
+    protected ISession Session;
+    protected readonly IMapper Mapper;
+
+    public RepositoryBase(ISession session)
     {
-        protected ISession Session;
-        protected IMapper Mapper;
-        
-        public RepositoryBase(ISession session)
-        {
-            Session = session;
-            Mapper = new Mapper(session);
-        }
+        Session = session;
+        Mapper = new Mapper(session);
     }
 }
