@@ -1,9 +1,5 @@
 ﻿using System;
 
-namespace plodo.Backend.Services.Exceptions
-{
-    public class SessionNotFoundException : Exception
-    {
-        
-    }
-}
+namespace plodo.Backend.Services.Exceptions;
+
+public class SessionNotFoundException : Exception;
