@@ -1,22 +1,19 @@
-using System;
+namespace plodo.Backend.API.Models;
 
-namespace plodo.Backend.API.Models
+public class CreateSessionResponse
 {
-    public class CreateSessionResponse
-    {
-        /// <summary>
-        /// The ID of the created session
-        /// </summary>
-        public string SessionId { get; set; }
+    /// <summary>
+    /// The ID of the created session
+    /// </summary>
+    public string SessionId { get; set; }
         
-        /// <summary>
-        /// The provided ID of the user hosting the session
-        /// </summary>
-        public Guid UserId { get; set; }
+    /// <summary>
+    /// The provided ID of the user hosting the session
+    /// </summary>
+    public Guid UserId { get; set; }
         
-        /// <summary>
-        /// Access token to be used in all subsequent requests. Should be placed in the Authorization header
-        /// </summary>
-        public AccessToken AccessToken { get; set; }
-    }
+    /// <summary>
+    /// Access token to be used in all subsequent requests. Should be placed in the Authorization header
+    /// </summary>
+    public AccessToken AccessToken { get; set; }
 }
